+++ examples/99 Ascent/src/Turret.h
@@ -38,7 +38,7 @@ private:
     void handle();
     iaVector3d getCurrentPos();
     void hitBy(uint64 entityID);
-    void onModelReady(uint32 modelNodeID);
+    void onModelReady(uint64 modelNodeID);
 
 };
